Backend/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
python-jwt
langchain
langchain-community
faiss-cpu
openai
python-docx
pypdf
python-dotenv